Implement input action duration
A simple mechanism is added to fetch the time in seconds to tell how
long an action has been pressed or released. The existing input actions
timestamps are used to calculate the duration between the current state
and the changed state similarly to `is_action_just_pressed/released`
methods the way it depends on physics/idle frame count.

Signed-off-by: bigmonte <34067397+bigmonte@users.noreply.github.com>